Improved data sheet organization
