+++ frontend/packages/sheet/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./provider";
 export * from "./components";
 export * from "./helpers";
-export * from "./columns";
 export * from "./virtualized";
